appveyor: add initial configuration file
