+++ test/test.go
@@ -1,0 +1,1 @@
+package kontrolclient
